chore: Spelling, Heading, Pagination btn and URL updated (#1782)
* Spelling, Heading and Link updated(Offical Docs)

* Pagination btn is now centred

* Relative in place of absolute

Absolute isn't the best to use in this case, the btn will never be centred, even with specified positions (left, top). Relative is best as it does not depend on that (top, right etc), plus, since no z-index will make the btn stack behind the footer, we need relative to keep in centred and allow for the use of the z-index to stack it on top of the footer for some screen resolutions.